Split up views into several different files
Also
 - added type annotations
 - added handling for when forms aren't valid, which was missing in a
   few cases
 - replaced HttpResponseRedirect(reverse(...))s with redirect(...)s
