+++ lrc_database/main/templatetags/groups.py
@@ -1,34 +1,36 @@
 from django import template
 
+from ..models import LRCDatabaseUser
+
 register = template.Library()
 
 
-def is_in_groups(user, *groups):
+def is_in_groups(user: LRCDatabaseUser, *groups: str) -> bool:
     return user.is_authenticated and user.groups.filter(name__in=groups).exists()
 
 
 @register.filter
-def is_si(user):
+def is_si(user: LRCDatabaseUser) -> bool:
     return is_in_groups(user, "SIs")
 
 
 @register.filter
-def is_tutor(user):
+def is_tutor(user: LRCDatabaseUser) -> bool:
     return is_in_groups(user, "Tutors")
 
 
 @register.filter
-def is_office_staff(user):
+def is_office_staff(user: LRCDatabaseUser) -> bool:
     return is_in_groups(user, "Office staff")
 
 
 @register.filter
-def is_supervisor(user):
+def is_supervisor(user: LRCDatabaseUser) -> bool:
     return is_in_groups(user, "Supervisors")
 
 
 @register.filter
-def is_privileged(user):
+def is_privileged(user: LRCDatabaseUser) -> bool:
     """
     Is the user someone who should have privileges to, say, view all users or
     add shifts to someone else? Currently, this just asks whether the user is
